Harden post-review: SQL cast, model labels, kiosk resilience, doc + copy re-sync.
Backend correctness:
- Cast retrieval_search semantic_trip_search limit to ::integer (the Phase
  4/5 graph-node path reintroduced the bigint-overload bug the search_agent
  path already fixed).
- bedrock_model_label: map the live fallback chain (Opus 4.8 -> Sonnet 4.6
  -> Haiku 4.5); drop the stale opus-4-7 entry.
- memory_server docstring: point at backend.memory.store, not the deleted
  memory_mcp_client.

Dead code removal:
- Delete orphaned memory_mcp_client.py (memory moved to backend.memory.store).
- Delete unmounted MeridianProChalkTalk.tsx + its CSS (stale Phase-N labels,
  zero importers).

Story-arc + copy:
- Re-sync talking-points doc to the live showcase: chorus = "a romantic slow
  week somewhere with great wine", Alex Morgan solo persona, real seeded
  facts, both demo hinges (Retrieval recall fails -> Production recalls;
  multi-intent Tokyo plan -> Workflow), Opus 4.8, Cohere Rerank 3.5.
- Fix stale marketing copy: "two founders" -> "small team", drop the old
  "romantic week in Europe" SQL example.

Kiosk (unattended booth):
- Graceful reconnect: replace raw fetch errors with a calm "Reconnecting to
  Aurora..." pill + auto-retry every 4s; Architecture/Try tabs stay usable.
- Re-sync demo scenarios to the showcase Tokyo storyline (beach -> tokyo):
  Tokyo seed -> recall -> multi-intent plan, matching Phase 4 pills exactly.

Verified: tsc clean, 44/44 tests pass.

+++ meridian/backend/routers/chat.py
@@ -1008,8 +1008,12 @@ async def retrieval_search(query: str, limit: int = 5) -> tuple[List[Product], L
         agent_file="agents/retrieval_03/search_agent.py"
     ))
     candidate_limit = max(limit * config.search.rerank_candidate_multiplier, 25)
+    # Cast the limit to ::integer — the function signature is
+    # semantic_trip_search(vector, integer); Python ints arrive over the
+    # RDS Data API as bigint and Postgres can't resolve the overload
+    # otherwise ("function semantic_trip_search(vector, bigint) does not exist").
     semantic_sql = """
-        SELECT * FROM semantic_trip_search(%s::vector, %s)
+        SELECT * FROM semantic_trip_search(%s::vector, %s::integer)
     """
     semantic_rows = await db.execute(semantic_sql, (embedding_str, candidate_limit))
     if price_filter is not None:
